// frontend/contracts/observer/v1/models/SubmissionsEventType.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export enum SubmissionsEventType {
    SUBMISSIONS_CREATED = 'submissions.created',
    SUBMISSIONS_QUEUED = 'submissions.queued',
    SUBMISSIONS_COMPILING_STARTED = 'submissions.compiling_started',
    SUBMISSIONS_TESTING_STARTED = 'submissions.testing_started',
    SUBMISSIONS_TEST_STARTED = 'submissions.test_started',
    SUBMISSIONS_COMPLETED = 'submissions.completed',
}